deliverable1/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cycler==0.11.0
grobid-client-python==0.0.4
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.4.3
nltk==3.8.1
Pillow==9.4.0
pip==22.3.1
pyparsing==3.0.9
python-dateutil==2.8.2
regex==2022.10.31
requests==2.28.2
setuptools==65.6.3
six==1.16.0
tornado==6.2
tqdm==4.64.1
urllib3==1.26.14
wheel==0.37.1
wordcloud==1.8.2.2
